Update patch set 1
Patch Set 1:

This is the new refactored/restructued ricplt deployment scripts and helm charts. It only works with the new ric-common-3.0.0 common template. This repo is standalone and can be deployed with the prerequisite of having ric-common-3.0.0 in the local helm repo. Such ric-common-3.0.0 should be installed by the it/dep repo scripts.

Patch-set: 1
Label: Code-Review=0