Update patch set 1
Patch Set 1:

Do we want to delay this until we fully test the dbaasha? otherwise now it will be deployed together when we deploy RIC, and it is not doing anything.

Also do we want to combine the dbaas_sa charts to dbaas_ha charts? It may be neater if we can just choose one using a recipe override.

Patch-set: 1
Reviewer: Gerrit User 1000046 <1000046@84e4e190-e623-4c37-9eda-58b1937c1b53>