Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> Not quite sure I understand the purpose here

The set -eu terminates the script when it hits a failed helm lint. The cat output will never be carried out in that case. As a result error message is not shown. Previous the output of helm lint is redirected to the output file because we run the script using docker build and we use the output file to scan for the error message. Now we prefer to directly print the output on screen instead of redirecting. It will shows tons of information more and stop right where the error is.

Patch-set: 2