Update patch set 6
Patch Set 6:

(3 comments)

By the look of this it will work with the new version of dbaas charts. The xapp-onboarder charts also need to be changed for this. Let me know if you want me to push a patch for xapp-onboarder.

Patch-set: 6
Reviewer: Gerrit User 1000046 <1000046@84e4e190-e623-4c37-9eda-58b1937c1b53>