Fix DBAAS helm charts timeout flag
DBAAS 0.5.0 and 0.5.1 cannot be deployed with default version of helm chart
what are under ric-plt/ric-dep repo. Reason is that for
ReadinessProbe/LivenessProbe checks there are used Redis ping command what is
shielded with timeout command to timeout in case ping happens to hang. DBAAS
docker images 0.5.0 and 0.5.1 are based on the Alpine 3.11 OS and in there
timeout command syntax has been changed not to have anymore '-t' flag. This
means that timeout '-t' flag cannot be used anymore in DBAAS helm charts for
the ReadinessProbe/LivenessProbe commands if installed DBAAS image is 0.5.0 or
newer one.

To fix this issue introduced new configuration option 'probeTimeoutCommand'
into values.yaml file what defines the timeout command syntax, by default its
value is "timeout" but it can be changed to "timeout -t" in case DBAAS image to
be installed is older than 0.5.0.

Issue-ID: RIC-761

Signed-off-by: Timo Tietavainen <timo.tietavainen@nokia.com>
Change-Id: Iecf599a02c69adfb9f821271310dceb682d4dbc7