Update patch set 4
Patch Set 4:

(6 comments)

> Patch Set 4:
> 
> is there anything we can do related to testing of this helm chart? It's a large commit, so some testing would be useful. This is related to RC-4 from here: https://wiki.o-ran-sc.org/display/RICP/Release+criteria+checklist+template

We do not have any reference related to testing in helm folder. This has been tested in our lab server and deploy-ric-platform  script is able to deploy the influxdb along with other RIC platform components. Please let me know if there is any specific testing we need to perform on this.

Patch-set: 4